implemented suggested changes from initial pr
